requirements.txt
Flask
gunicorn
flask-cors
tensorflow-cpu
transformers
numpy
json
gdown==4.1.1
